+++ cli/src/api/index.ts
@@ -1,4 +1,0 @@
-export * from './api'
-export * from './types'
-export * from './deploys'
-export * from './squids'