+++ substrate-processor/src/batch/request.ts
@@ -1,5 +1,5 @@
-import {EvmTopicSet} from "../interfaces/dataHandlers"
-import {CallDataRequest, EventDataRequest} from "../interfaces/dataSelection"
+import {EvmTopicSet} from '../interfaces/dataHandlers'
+import {CallDataRequest, EventDataRequest} from '../interfaces/dataSelection'
 
 
 type EventReq = {
@@ -21,19 +21,26 @@ type EvmLogReq = {
 }
 
 
-type ContractsEventsReq = {
+type EthereumTransactionReq = {
+    contract: string
+    sighash?: string
+    data?: CallDataRequest
+}
+
+
+type ContractsEventReq = {
     contract: string
     data?: EventDataRequest
 }
 
 
-type GearMessagesEnqueuedReq = {
+type GearMessageEnqueuedReq = {
     program: string
     data?: EventDataRequest
 }
 
 
-type GearUserMessagesSentReq = {
+type GearUserMessageSentReq = {
     program: string
     data?: EventDataRequest
 }
@@ -65,9 +72,10 @@ export interface BatchRequest {
     getEvents(): EventReq[]
     getCalls(): CallReq[]
     getEvmLogs(): EvmLogReq[]
-    getContractsEvents(): ContractsEventsReq[]
-    getGearMessagesEnqueued(): GearMessagesEnqueuedReq[]
-    getGearUserMessagesSent(): GearUserMessagesSentReq[]
+    getEthereumTransactions(): EthereumTransactionReq[]
+    getContractsEvents(): ContractsEventReq[]
+    getGearMessagesEnqueued(): GearMessageEnqueuedReq[]
+    getGearUserMessagesSent(): GearUserMessageSentReq[]
     getAcalaEvmExecuted(): AcalaEvmExecutedReq[]
     getAcalaEvmCall(): AcalaEvmCallReq[]
     getAcalaEvmEthCall(): AcalaEvmEthCallReq[]
@@ -78,9 +86,10 @@ export class PlainBatchRequest implements BatchRequest {
     events: EventReq[] = []
     calls: CallReq[] = []
     evmLogs: EvmLogReq[] = []
-    contractsEvents: ContractsEventsReq[] = []
-    gearMessagesEnqueued: GearMessagesEnqueuedReq[] = []
-    gearUserMessagesSent: GearUserMessagesSentReq[] = []
+    ethereumTransactions: EthereumTransactionReq[] = []
+    contractsEvents: ContractsEventReq[] = []
+    gearMessagesEnqueued: GearMessageEnqueuedReq[] = []
+    gearUserMessagesSent: GearUserMessageSentReq[] = []
     acalaEvmExecuted: AcalaEvmExecutedReq[] = []
     acalaEvmCall: AcalaEvmCallReq[] = []
     acalaEvmEthCall: AcalaEvmEthCallReq[] = []
@@ -98,15 +107,19 @@ export class PlainBatchRequest implements BatchRequest {
         return this.evmLogs
     }
 
-    getContractsEvents(): ContractsEventsReq[] {
+    getEthereumTransactions(): EthereumTransactionReq[] {
+        return this.ethereumTransactions
+    }
+
+    getContractsEvents(): ContractsEventReq[] {
         return this.contractsEvents
     }
 
-    getGearMessagesEnqueued(): GearMessagesEnqueuedReq[] {
+    getGearMessagesEnqueued(): GearMessageEnqueuedReq[] {
         return this.gearMessagesEnqueued
     }
 
-    getGearUserMessagesSent(): GearUserMessagesSentReq[] {
+    getGearUserMessagesSent(): GearUserMessageSentReq[] {
         return this.gearUserMessagesSent
     }
 
@@ -132,6 +145,7 @@ export class PlainBatchRequest implements BatchRequest {
         result.events = this.events.concat(other.events)
         result.calls = this.calls.concat(other.calls)
         result.evmLogs = this.evmLogs.concat(other.evmLogs)
+        result.ethereumTransactions = this.ethereumTransactions.concat(other.ethereumTransactions)
         result.contractsEvents = this.contractsEvents.concat(other.contractsEvents)
         result.gearMessagesEnqueued = this.gearMessagesEnqueued.concat(other.gearMessagesEnqueued)
         result.gearUserMessagesSent = this.gearUserMessagesSent.concat(other.gearUserMessagesSent)