+++ substrate-processor/src/processor/handlerProcessor.ts
@@ -1,11 +1,12 @@
-import {createLogger, Logger} from "@subsquid/logger"
-import {getOldTypesBundle, OldTypesBundle, QualifiedName, readOldTypesBundle} from "@subsquid/substrate-metadata"
-import {assertNotNull, def, runProgram, unexpectedCase} from "@subsquid/util-internal"
-import assert from "assert"
-import {applyRangeBound, Batch, mergeBatches} from "../batch/generic"
-import {CallHandlerEntry, DataHandlers} from "../batch/handlers"
-import type {Chain} from "../chain"
-import type {BlockData} from "../ingest"
+import {createLogger, Logger} from '@subsquid/logger'
+import {getOldTypesBundle, OldTypesBundle, QualifiedName, readOldTypesBundle} from '@subsquid/substrate-metadata'
+import {assertNotNull, def, runProgram, unexpectedCase} from '@subsquid/util-internal'
+import assert from 'assert'
+import {applyRangeBound, Batch, mergeBatches} from '../batch/generic'
+import {CallHandlerEntry, DataHandlers} from '../batch/handlers'
+import type {Chain} from '../chain'
+import type {BlockData} from '../ingest'
+import {EthereumTransactionHandlerOptions} from '../interfaces/dataHandlers'
 import type {
     BlockHandler,
     BlockHandlerContext,
@@ -22,32 +23,30 @@ import type {
     GearMessageEnqueuedHandler,
     GearUserMessageSentHandler,
     AcalaEvmExecutedHandler,
-    AcalaEvmCallHandler,
-    AcalaEvmEthCallHandler,
     AcalaEvmExecutedOptions,
     AcalaEvmCallHandlerOptions,
-    AcalaEvmEthCallHandlerOptions,
-} from "../interfaces/dataHandlers"
+    AcalaEvmEthCallHandlerOptions
+} from '../interfaces/dataHandlers'
 import type {
     CallDataRequest,
     DataSelection,
     EventDataRequest,
     MayBeDataSelection,
     NoDataSelection
-} from "../interfaces/dataSelection"
-import type {Database} from "../interfaces/db"
-import type {Hooks} from "../interfaces/hooks"
+} from '../interfaces/dataSelection'
+import type {Database} from '../interfaces/db'
+import type {Hooks} from '../interfaces/hooks'
 import type {
     ContractsContractEmittedEvent,
+    EvmLogEvent,
     GearMessageEnqueuedEvent,
     GearUserMessageSentEvent,
-    EvmLogEvent,
     SubstrateCall,
-    SubstrateEvent,
-} from "../interfaces/substrate"
-import {withErrorContext} from "../util/misc"
-import type {Range} from "../util/range"
-import {Options, Runner} from "./runner"
+    SubstrateEvent
+} from '../interfaces/substrate'
+import {withErrorContext} from '../util/misc'
+import type {Range} from '../util/range'
+import {Options, Runner} from './runner'
 
 
 export interface DataSource {
@@ -72,6 +71,7 @@ export class SubstrateProcessor<Store> {
         event: [],
         call: [],
         evmLog: [],
+        ethereumTransaction: [],
         contractsContractEmitted: [],
         gearMessageEnqueued: [],
         gearUserMessageSent: [],
@@ -260,7 +260,7 @@ export class SubstrateProcessor<Store> {
     addPreHook(fn: BlockHandler<Store>): this
     addPreHook(options: BlockRangeOption & NoDataSelection, fn: BlockHandler<Store>): this
     addPreHook<R extends BlockHandlerDataRequest>(options: BlockRangeOption & DataSelection<R>, fn: BlockHandler<Store, R>): this
-    addPreHook(fnOrOptions: BlockHandler<Store> | BlockRangeOption & MayBeDataSelection<BlockHandlerDataRequest> , fn?: BlockHandler<Store>): this {
+    addPreHook(fnOrOptions: BlockHandler<Store> | BlockRangeOption & MayBeDataSelection<BlockHandlerDataRequest>, fn?: BlockHandler<Store>): this {
         this.assertNotRunning()
         let handler: BlockHandler<Store>
         let options: BlockRangeOption & MayBeDataSelection<BlockHandlerDataRequest> = {}
@@ -385,7 +385,7 @@ export class SubstrateProcessor<Store> {
      */
     addEventHandler(eventName: QualifiedName, fn: EventHandler<Store>): this
     addEventHandler(eventName: QualifiedName, options: BlockRangeOption & NoDataSelection, fn: EventHandler<Store>): this
-    addEventHandler<R extends EventDataRequest>(eventName: QualifiedName, options: BlockRangeOption & DataSelection<R>, fn: EventHandler<Store, R> ): this
+    addEventHandler<R extends EventDataRequest>(eventName: QualifiedName, options: BlockRangeOption & DataSelection<R>, fn: EventHandler<Store, R>): this
     addEventHandler(eventName: QualifiedName, fnOrOptions: BlockRangeOption & MayBeDataSelection<EventDataRequest> | EventHandler<Store>, fn?: EventHandler<Store>): this {
         this.assertNotRunning()
         let handler: EventHandler<Store>
@@ -446,7 +446,7 @@ export class SubstrateProcessor<Store> {
     addCallHandler(callName: QualifiedName, fnOrOptions: CallHandler<Store> | CallHandlerOptions & MayBeDataSelection<CallDataRequest>, fn?: CallHandler<Store>): this {
         this.assertNotRunning()
         let handler: CallHandler<Store>
-        let options:  CallHandlerOptions & MayBeDataSelection<CallDataRequest> = {}
+        let options: CallHandlerOptions & MayBeDataSelection<CallDataRequest> = {}
         if (typeof fnOrOptions == 'function') {
             handler = fnOrOptions
         } else {
@@ -465,48 +465,120 @@ export class SubstrateProcessor<Store> {
      * Registers `EVM.Log` event handler.
      *
      * This method is similar to {@link .addEventHandler},
-     * but provides specialised {@link EvmLogEvent | event type} and selects
-     * events by evm log contract address and topics.
+     * but selects events by evm log contract address and topics.
      *
      * @example
      * // process ERC721 transfers from Moonsama contract
      * processor.addEvmLogHandler('0xb654611f84a8dc429ba3cb4fda9fad236c505a1a', {
      *     topics: ['0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef']
      * }, async ctx => {})
+     *
+     * // subscribe to multiple contracts at once
+     * processor.addEvmLogHandler([
+     *     '0xb654611f84a8dc429ba3cb4fda9fad236c505a1a',
+     *     '0x6a2d262D56735DbA19Dd70682B39F6bE9a931D98'
+     * ], {
+     *     topics: ['0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef']
+     * }, async ctx => {})
      */
     addEvmLogHandler(
-        contractAddress: string,
+        contractAddress: string | string[],
         fn: EvmLogHandler<Store>
     ): this
     addEvmLogHandler(
-        contractAddress: string,
+        contractAddress: string | string[],
         options: EvmLogOptions & NoDataSelection,
         fn: EvmLogHandler<Store>
     ): this
     addEvmLogHandler<R extends EventDataRequest>(
-        contractAddress: string,
+        contractAddress: string | string[],
         options: EvmLogOptions & DataSelection<R>,
         fn: EvmLogHandler<Store, R>
     ): this
     addEvmLogHandler(
-        contractAddress: string,
+        contractAddress: string | string[],
         fnOrOptions: EvmLogOptions & MayBeDataSelection<EventDataRequest> | EvmLogHandler<Store>,
         fn?: EvmLogHandler<Store>
     ): this {
         this.assertNotRunning()
         let handler: EvmLogHandler<Store>
-        let options:  EvmLogOptions= {}
+        let options: EvmLogOptions = {}
         if (typeof fnOrOptions == 'function') {
             handler = fnOrOptions
         } else {
             handler = assertNotNull(fn)
             options = {...fnOrOptions}
         }
-        this.hooks.evmLog.push({
+        let contractAddresses = Array.isArray(contractAddress) ? contractAddress : [contractAddress]
+        this.hooks.evmLog.push(...contractAddresses.map((contractAddress) => ({
             handler,
             contractAddress: contractAddress.toLowerCase(),
             ...options
-        })
+        })))
+        return this
+    }
+
+    /**
+     * Registers `Ethereum.transact` call handler.
+     *
+     * This method is similar to {@link .addCallHandler},
+     * but selects calls by contract address and sighash.
+     * Only "EVM call" transactions are selected.
+     *
+     * Note, that there is a difference between success of substrate call
+     * and successful execution of its EVM transaction.
+     * {@link EthereumTransactionHandlerOptions.triggerForFailedCalls}
+     * option refers to success of substrate call.
+     *
+     * @example
+     * // process EVM calls to contract `0x6a2d262D56735DbA19Dd70682B39F6bE9a931D98`
+     * process.addEthereumTransactionHandler('0x6a2d262D56735DbA19Dd70682B39F6bE9a931D98', async ctx => {})
+     *
+     * // process all EVM calls with signature `transfer(address,uint256)`
+     * process.addEthereumTransactionHandler('*', {sighash: '0xa9059cbb'}, async ctx => {})
+     *
+     * // subscribe to multiple contracts at once
+     * processor.addEthereumTransaction([
+     *     '0x6a2d262D56735DbA19Dd70682B39F6bE9a931D98',
+     *     '0x3795C36e7D12A8c252A20C5a7B455f7c57b60283'
+     * ], {
+     *     sighash: '0xa9059cbb'
+     * }, async ctx => {})
+     */
+    addEthereumTransactionHandler(
+        contractAddress: string | string[],
+        fn: CallHandler<Store>
+    ): this
+    addEthereumTransactionHandler(
+        contractAddress: string | string[],
+        options: EthereumTransactionHandlerOptions & NoDataSelection,
+        fn: CallHandler<Store>
+    ): this
+    addEthereumTransactionHandler<R extends CallDataRequest>(
+        contractAddress: string | string[],
+        options: EthereumTransactionHandlerOptions & DataSelection<R>,
+        fn: CallHandler<Store, R>
+    ): this
+    addEthereumTransactionHandler(
+        contractAddress: string | string[],
+        fnOrOptions: EthereumTransactionHandlerOptions & MayBeDataSelection<CallDataRequest> | CallHandler<Store>,
+        fn?: CallHandler<Store>
+    ): this {
+        this.assertNotRunning()
+        let handler: CallHandler<Store>
+        let options: EthereumTransactionHandlerOptions = {}
+        if (typeof fnOrOptions == 'function') {
+            handler = fnOrOptions
+        } else {
+            handler = assertNotNull(fn)
+            options = {...fnOrOptions}
+        }
+        let contractAddresses = Array.isArray(contractAddress) ? contractAddress : [contractAddress]
+        this.hooks.ethereumTransaction.push(...contractAddresses.map((contractAddress) => ({
+            handler,
+            contractAddress: contractAddress.toLowerCase(),
+            ...options
+        })))
         return this
     }
 
@@ -621,7 +693,7 @@ export class SubstrateProcessor<Store> {
         programId: string,
         fnOrOptions: GearUserMessageSentHandler<Store> | BlockRangeOption & MayBeDataSelection<EventDataRequest>,
         fn?: GearUserMessageSentHandler<Store>
-    ): this{
+    ): this {
         this.assertNotRunning()
         let handler: GearUserMessageSentHandler<Store>
         let options: BlockRangeOption & MayBeDataSelection<EventDataRequest> = {}
@@ -704,25 +776,25 @@ export class SubstrateProcessor<Store> {
      */
     addAcalaEvmCallHandler(
         contractAddress: string,
-        fn: AcalaEvmCallHandler<Store>
+        fn: CallHandler<Store>
     ): this
     addAcalaEvmCallHandler(
         contractAddress: string,
         options: AcalaEvmCallHandlerOptions & NoDataSelection,
-        fn: AcalaEvmCallHandler<Store>
+        fn: CallHandler<Store>
     ): this
     addAcalaEvmCallHandler<R extends CallDataRequest>(
         contractAddress: string,
         options: AcalaEvmCallHandlerOptions & DataSelection<R>,
-        fn: AcalaEvmCallHandler<Store, R>
+        fn: CallHandler<Store, R>
     ): this
     addAcalaEvmCallHandler(
         contractAddress: string,
-        fnOrOptions: AcalaEvmCallHandlerOptions & MayBeDataSelection<CallDataRequest> | AcalaEvmCallHandler<Store>,
-        fn?: AcalaEvmCallHandler<Store>
+        fnOrOptions: AcalaEvmCallHandlerOptions & MayBeDataSelection<CallDataRequest> | CallHandler<Store>,
+        fn?: CallHandler<Store>
     ): this {
         this.assertNotRunning()
-        let handler: AcalaEvmCallHandler<Store>
+        let handler: CallHandler<Store>
         let options: AcalaEvmCallHandlerOptions = {}
         if (typeof fnOrOptions == 'function') {
             handler = fnOrOptions
@@ -753,25 +825,25 @@ export class SubstrateProcessor<Store> {
      */
     addAcalaEvmEthCallHandler(
         contractAddress: string,
-        fn: AcalaEvmEthCallHandler<Store>
+        fn: CallHandler<Store>
     ): this
     addAcalaEvmEthCallHandler(
         contractAddress: string,
         options: AcalaEvmEthCallHandlerOptions & NoDataSelection,
-        fn: AcalaEvmEthCallHandler<Store>
+        fn: CallHandler<Store>
     ): this
     addAcalaEvmEthCallHandler<R extends CallDataRequest>(
         contractAddress: string,
         options: AcalaEvmEthCallHandlerOptions & DataSelection<R>,
-        fn: AcalaEvmEthCallHandler<Store, R>
+        fn: CallHandler<Store, R>
     ): this
     addAcalaEvmEthCallHandler(
         contractAddress: string,
-        fnOrOptions: AcalaEvmEthCallHandlerOptions & MayBeDataSelection<CallDataRequest> | AcalaEvmEthCallHandler<Store>,
-        fn?: AcalaEvmEthCallHandler<Store>
+        fnOrOptions: AcalaEvmEthCallHandlerOptions & MayBeDataSelection<CallDataRequest> | CallHandler<Store>,
+        fn?: CallHandler<Store>
     ): this {
         this.assertNotRunning()
-        let handler: AcalaEvmEthCallHandler<Store>
+        let handler: CallHandler<Store>
         let options: AcalaEvmEthCallHandlerOptions = {}
         if (typeof fnOrOptions == 'function') {
             handler = fnOrOptions
@@ -796,7 +868,7 @@ export class SubstrateProcessor<Store> {
     private createBatches(blockRange: Range) {
         let batches: Batch<DataHandlers>[] = []
 
-        function getRange(hook: { range?: Range }): Range{
+        function getRange(hook: {range?: Range}): Range {
             return hook.range || {from: 0}
         }
 
@@ -846,6 +918,23 @@ export class SubstrateProcessor<Store> {
                     filter: hook.filter,
                     handler: hook.handler
                 }]
+            }
+            batches.push({range, request})
+        })
+
+        this.hooks.ethereumTransaction.forEach(hook => {
+            let range = getRange(hook)
+            let request = new DataHandlers()
+            request.ethereumTransactions = {
+                [hook.contractAddress]: {
+                    [hook.sighash || '*']: {
+                        data: hook.data,
+                        handlers: [{
+                            handler: hook.handler,
+                            triggerForFailedCalls: hook.triggerForFailedCalls
+                        }]
+                    }
+                }
             }
             batches.push({range, request})
         })
@@ -1029,7 +1118,7 @@ class HandlerRunner<S> extends Runner<S, DataHandlers>{
         }
 
         for (let item of block.items) {
-            switch(item.kind) {
+            switch (item.kind) {
                 case 'event':
                     for (let handler of this.getEventHandlers(handlers, item.event)) {
                         let log = blockLog.child({
@@ -1130,6 +1219,19 @@ class HandlerRunner<S> extends Runner<S, DataHandlers>{
                             log.debug('end')
                         }
                     }
+                    for (let handler of this.getEthereumTransactionHandlers(handlers, item.call)) {
+                        if (item.call.success || handler.triggerForFailedCalls) {
+                            let log = blockLog.child({
+                                hook: 'ethereum-transaction',
+                                callName: item.call.name,
+                                callId: item.call.id
+                            })
+                            let {kind, ...data} = item
+                            log.debug('begin')
+                            await handler.handler({...ctx, log, ...data})
+                            log.debug('end')
+                        }
+                    }
                     for (let handler of this.getAcalaEvmCallHandlers(handlers, item.call)) {
                         if (item.call.success || handler.triggerForFailedCalls) {
                             let log = blockLog.child({
@@ -1172,26 +1274,32 @@ class HandlerRunner<S> extends Runner<S, DataHandlers>{
     }
 
     private *getEventHandlers(handlers: DataHandlers, event: SubstrateEvent): Generator<EventHandler<any>, any, any> {
-        let hs = handlers.events['*']
-        if (hs) {
-            yield* hs.handlers
-        }
-        hs = handlers.events[event.name]
-        if (hs) {
+        for (let hs of extract(handlers.events, event.name)) {
             yield* hs.handlers
         }
     }
 
     private *getCallHandlers(handlers: DataHandlers, call: SubstrateCall): Generator<CallHandlerEntry, any, any> {
-        let hs = handlers.calls['*']
-        if (hs) {
-            yield* hs.handlers
-        }
-        hs = handlers.calls[call.name]
-        if (hs) {
-            yield* hs.handlers
+       for (let hs of extract(handlers.calls, call.name)) {
+           yield* hs.handlers
+       }
+    }
+
+    private *getEthereumTransactionHandlers(handlers: DataHandlers, call: SubstrateCall): Generator<CallHandlerEntry, any, any> {
+        if (call.name != 'Ethereum.transact') return
+        let tx = call.args.transaction.value.action ? call.args.transaction.value : call.args.transaction
+
+        let contractAddress = tx.action.value
+        if (contractAddress == null) return
+        let sighash = tx.input.slice(0, 10)
+
+        for (let sighashMap of extract(handlers.ethereumTransactions, contractAddress)) {
+            for (let hs of extract(sighashMap, sighash)) {
+                yield* hs.handlers
+            }
         }
     }
+
 
     private *getEvmLogHandlers(evmLogs: DataHandlers["evmLogs"], event: SubstrateEvent): Generator<EvmLogHandler<any>> {
         if (event.name != 'EVM.Log') return