+++ stores/typeorm-store/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @subsquid/typeorm-store
 
-This log was last generated on Thu, 07 Jul 2022 15:00:50 GMT and should not be manually modified.
+This log was last generated on Wed, 17 Aug 2022 17:33:52 GMT and should not be manually modified.
+
+## 0.1.2
+Wed, 17 Aug 2022 17:33:52 GMT
+
+_Version update only_
 
 ## 0.1.1
 Thu, 07 Jul 2022 15:00:50 GMT